+++ lib/parser/base.dart
@@ -5,6 +5,7 @@ mixin MetadataKeys {
   static const kDescription = 'description';
   static const kImage = 'image';
   static const kUrl = 'url';
+  static const kTimeout = 'timeout';
 }
 
 mixin BaseMetaInfo {
@@ -30,7 +31,7 @@ abstract class InfoBase {
 /// Container class for Metadata
 class Metadata extends InfoBase with BaseMetaInfo, MetadataKeys {
   bool get hasAllMetadata {
-    return (title != null && desc != null && image != null && url != null);
+    return title != null && desc != null && image != null && url != null;
   }
 
   @override
@@ -38,12 +39,13 @@ class Metadata extends InfoBase with BaseMetaInfo, MetadataKeys {
     return toMap().toString();
   }
 
-  Map<String, String?> toMap() {
+  Map<String, dynamic> toMap() {
     return {
       MetadataKeys.kTitle: title,
       MetadataKeys.kDescription: desc,
       MetadataKeys.kImage: image,
       MetadataKeys.kUrl: url,
+      MetadataKeys.kTimeout: timeout.millisecondsSinceEpoch ~/ 1000,
     };
   }
 
@@ -57,6 +59,8 @@ class Metadata extends InfoBase with BaseMetaInfo, MetadataKeys {
     m.desc = json[MetadataKeys.kDescription];
     m.image = json[MetadataKeys.kImage];
     m.url = json[MetadataKeys.kUrl];
+    m.timeout = DateTime.fromMillisecondsSinceEpoch(
+        json[MetadataKeys.kTimeout]! * 1000);
     return m;
   }
 }