+++ lib/ui/link_view_vertical.dart
@@ -6,7 +6,7 @@ class LinkViewVertical extends StatelessWidget {
   final String title;
   final String description;
   final String imageUri;
-  final Function(String) onTap;
+  final Function() onTap;
   final TextStyle? titleTextStyle;
   final TextStyle? bodyTextStyle;
   final bool? showMultiMedia;
@@ -32,7 +32,7 @@ class LinkViewVertical extends StatelessWidget {
   }) : super(key: key);
 
   double computeTitleFontSize(double height) {
-    double size = height * 0.13;
+    var size = height * 0.13;
     if (size > 15) {
       size = 15;
     }
@@ -66,7 +66,7 @@ class LinkViewVertical extends StatelessWidget {
             fontWeight: FontWeight.w400,
           );
 
-      ImageProvider? _img = imageUri != "" ? NetworkImage(imageUri) : null;
+      ImageProvider? _img = imageUri != '' ? NetworkImage(imageUri) : null;
       if (imageUri.startsWith('data:image')) {
         _img = MemoryImage(
           base64Decode(imageUri.substring(imageUri.indexOf('base64') + 7)),
@@ -74,7 +74,7 @@ class LinkViewVertical extends StatelessWidget {
       }
 
       return InkWell(
-          onTap: () => onTap(url),
+          onTap: () => onTap(),
           child: Column(
             children: <Widget>[
               showMultiMedia!